% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.scope.R
\name{predict.scope}
\alias{predict.scope}
\title{Computes SCOPE predictions}
\usage{
\method{predict}{scope}(object, newdata, interceptxlinear = FALSE, ...)
}
\arguments{
\item{object}{SCOPE model as outputted by scope. Must have simply.the.best = TRUE}

\item{newdata}{New covariates on which to make predictions. Must be of the same form as the model was trained on}

\item{interceptxlinear}{If intercept is to be included in the model and is not in the column span of the continuous variables in x, set to FALSE (default).
Must match format of training data.}

\item{...}{Additional arguments to pass to other \code{predict} methods}
}
\value{
Returns n-length vector of predictions
}
\description{
Computes SCOPE predictions on new data.
}
\seealso{
\code{\link{scope}}
}
