% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniformDesignMatrix.R
\name{UniformDesignMatrix}
\alias{UniformDesignMatrix}
\title{Create a design matrix of categorical variables.}
\usage{
UniformDesignMatrix(n, p, c)
}
\arguments{
\item{n}{Number of observations}

\item{p}{Number of variables}

\item{c}{Number of categories within each variable}
}
\value{
A data frame of categorical (factor) variables.
}
\description{
Function for use in simulations, creating design matrix of categorical variables where each column is uniformly randomly distributed and independent.
}
\examples{
x = UniformDesignMatrix(100, 10, 8)
}
